backend/requirements.txt
fastapi
uvicorn
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fuzzywuzzy
python-Levenshtein
apscheduler